HBASE-5654 [findbugs] Address dodgy bugs (Ashutosh Jindal)
git-svn-id: https://svn.apache.org/repos/asf/hbase/trunk@1329132 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/hadoop/hbase/io/hfile/LruBlockCache.java
@@ -671,7 +671,7 @@ public class LruBlockCache implements BlockCache, HeapSize {
     return CACHE_FIXED_OVERHEAD + ClassSize.CONCURRENT_HASHMAP +
         ((long)Math.ceil(maxSize*1.2/blockSize)
             * ClassSize.CONCURRENT_HASHMAP_ENTRY) +
-        (concurrency * ClassSize.CONCURRENT_HASHMAP_SEGMENT);
+        ((long)concurrency * ClassSize.CONCURRENT_HASHMAP_SEGMENT);
   }
 
   @Override
